pyproject.toml
[project]
name = "ops_ai_cortex"
version = "0.1.0"
description = "AI-driven DevOps intelligence platform"
readme = "README.md"
requires-python = ">=3.12"

dependencies = [
  "fastapi",
  "langchain",
  "redis",
  "psycopg2-binary",
  "anthropic",
  "openai",
  "pydantic",
  "uvicorn",
  "jose>=1.0.0",
  "python-jose[cryptography]>=3.5.0",
  "passlib[bcrypt]>=1.7.4",
  "fastapi-users-db-sqlalchemy>=7.0.0",
  "email-validator>=2.3.0",
]

[tool.uv]
dev-dependencies = [
  "pytest",
  "black",
  "ruff",
  "mypy",
  "httpx",
  "pytest-asyncio"
]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
